test: route cache in middleware/cache
